Share setup of datasource between tests
